Add confs, docs, wrp_client dirs in bin; basic_install in adios/docs. Update bin/readme; adios/readme; readme
