--- bin/old/requirements-wrp-chat.txt
@@ -1,2 +1,0 @@
-ollama>=0.1.0
-mcp>=0.1.0
