+++ bin/wrp_client/mcp_manager.py
@@ -1,0 +1,96 @@
+import asyncio
+import sys
+import os
+from contextlib import AsyncExitStack
+from pathlib import Path
+from typing import List, Optional
+
+from mcp import ClientSession
+from mcp.client.stdio import StdioServerParameters, stdio_client
+
+from .providers.base import BaseLLM, ToolDef
+
+def find_server_py(server_name: str) -> str:
+    """
+    Locates the main server.py script for a given MCP server name.
+    """
+    repo_root = Path(__file__).resolve().parent.parent.parent
+    for search_path in [repo_root / server_name / "src", repo_root / server_name]:
+        if search_path.exists():
+            matches = list(search_path.rglob("server.py"))
+            if matches:
+                return str(matches[0])
+    raise FileNotFoundError(f"Could not find server.py for {server_name}")
+
+
+class MCPManager:
+    """
+    Manages the connection and interaction with a single MCP server.
+    """
+    def __init__(self, llm_adapter: BaseLLM):
+        self.session: Optional[ClientSession] = None
+        self.exit_stack = AsyncExitStack()
+        self.llm = llm_adapter
+        self.tools: List[ToolDef] = []
+
+    async def connect(self, server_script: str):
+        print(f"Starting server with stdio: {server_script}")
+        params = StdioServerParameters(command=sys.executable, args=[server_script], env=os.environ)
+        reader, writer = await self.exit_stack.enter_async_context(stdio_client(params))
+        self.session = await self.exit_stack.enter_async_context(ClientSession(reader, writer))
+        
+        await self.session.initialize()
+        tool_list = await self.session.list_tools()
+        
+        self.tools = [
+            ToolDef(
+                name=t.name,
+                description=t.description,
+                input_schema=t.inputSchema,
+            )
+            for t in tool_list.tools
+        ]
+        print("\nConnected. Tools available:")
+        for tool in self.tools:
+            print(f" * {tool.name}: {tool.description}")
+
+    async def process_query(self, query: str) -> str:
+        messages = [{"role": "user", "content": query}]
+        
+        try:
+            llm_reply = await self.llm.chat(messages, self.tools)
+            
+            if llm_reply.tool_calls:
+                final_parts = []
+                for tool_call in llm_reply.tool_calls:
+                    name, args = tool_call["name"], tool_call["args"]
+                    print(f"[Calling tool {name} with args {args}]")
+                    try:
+                        tr = await self.session.call_tool(name, args)
+                        raw_txt = tr.content[0].text if tr.content else "No content returned"
+                        final_parts.append(f"[Called {name}: {raw_txt}]")
+                    except Exception as e:
+                        final_parts.append(f"[Error calling {name}: {e}]")
+                return "\n".join(final_parts)
+            else:
+                return llm_reply.text
+
+        except Exception as e:
+            return f"Error during LLM processing: {e}"
+
+    async def chat_loop(self):
+        print(f"\nMCP Client Started! (type 'quit' to exit)")
+        try:
+            while True:
+                q = input("\nQuery: ").strip()
+                if q.lower() in ('quit', 'exit'):
+                    break
+                response_text = await self.process_query(q)
+                print(f"\n{response_text}")
+        except (KeyboardInterrupt, EOFError):
+            print("\nExiting...")
+        finally:
+            await self.cleanup()
+
+    async def cleanup(self):
+        await self.exit_stack.aclose()
